# src/main/resources/openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: Request for persons
  version: 1.0.0
  description: OpenAPI Specification for persons

servers:
  - url: /delphin

paths:
  /persons:
    get:
      tags:
        - Persons
      description: Получение списка пользователей
      operationId: get-all-persons
      responses:
        200:
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: 'model/persons_response.json'
        404:
          description: Not Found
        400:
          description: Bad Request
        401:
          description: Unauthorized
        403:
          description: Forbidden

  /persons/{personId}:
    get:
      tags:
        - Person
      description: Получение пользоввателя по id
      operationId: get-by-id
      parameters:
        - name: personId
          in: path
          required: true
          schema:
            type: string
          description: Идентификатор пользователя
      responses:
        200:
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: 'model/person_response.json'
        404:
          description: Not Found
        400:
          description: Bad Request
        401:
          description: Unauthorized
        403:
          description: Forbidden

  /persons/{personId}/tasks:
    get:
      tags:
        - All Tasks by PersonId
      description: Получение списка задач для пользователя
      operationId: get-all-tasks-by-person-id
      parameters:
        - name: personId
          in: path
          required: true
          schema:
            type: string
          description: Идентификатор пользователя
      responses:
        200:
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: 'model/tasks_response.json'
        404:
          description: Not Found
        400:
          description: Bad Request
        401:
          description: Unauthorized
        403:
          description: Forbidden

  /tasks/{taskId}:
    get:
      tags:
        - Task
      description: Получение задачи по id
      operationId: get-task-by-id
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
          description: Идентификатор задачи
      responses:
        200:
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: 'model/task_response.json'
        404:
          description: Not Found
        400:
          description: Bad Request
        401:
          description: Unauthorized
        403:
          description: Forbidden